feat(Compatibility) finish SchemaParserSpecification
