+++ spec/graphql/compatibility/query_parser_specification_spec.rb
@@ -1,5 +1,5 @@
 require "spec_helper"
 
-BuiltInParserSuite = GraphQL::Compatibility::QueryParserSpecification.build_suite do |query_string|
+BuiltInQueryParserSuite = GraphQL::Compatibility::QueryParserSpecification.build_suite do |query_string|
   GraphQL::Language::Parser.parse(query_string)
 end
